+++ delivery-service/src/main/java/com/sobok/deliveryservice/delivery/service/DeliveryService.java
@@ -2,23 +2,21 @@ package com.sobok.deliveryservice.delivery.service;
 
 import com.sobok.deliveryservice.common.dto.TokenUserInfo;
 import com.sobok.deliveryservice.common.exception.CustomException;
-import com.sobok.deliveryservice.delivery.client.AuthFeignClient;
 import com.sobok.deliveryservice.delivery.client.PaymentFeignClient;
 import com.sobok.deliveryservice.delivery.client.ShopFeignClient;
 import com.sobok.deliveryservice.delivery.client.UserFeignClient;
-import com.sobok.deliveryservice.delivery.dto.info.AuthRiderInfoResDto;
 import com.sobok.deliveryservice.delivery.dto.info.UserAddressDto;
-import com.sobok.deliveryservice.delivery.dto.payment.RiderChangeOrderStateReqDto;
 import com.sobok.deliveryservice.delivery.dto.payment.DeliveryRegisterDto;
 import com.sobok.deliveryservice.delivery.dto.payment.RiderPaymentInfoResDto;
 import com.sobok.deliveryservice.delivery.dto.payment.ShopPaymentResDto;
 import com.sobok.deliveryservice.delivery.dto.request.AcceptOrderReqDto;
-import com.sobok.deliveryservice.delivery.dto.request.RiderReqDto;
-import com.sobok.deliveryservice.delivery.dto.response.*;
+import com.sobok.deliveryservice.delivery.dto.response.DeliveryAvailOrderResDto;
+import com.sobok.deliveryservice.delivery.dto.response.DeliveryAvailShopResDto;
+import com.sobok.deliveryservice.delivery.dto.response.DeliveryOrderResDto;
+import com.sobok.deliveryservice.delivery.dto.response.DeliveryResDto;
 import com.sobok.deliveryservice.delivery.entity.Delivery;
 import com.sobok.deliveryservice.delivery.entity.Rider;
 import com.sobok.deliveryservice.delivery.repository.DeliveryRepository;
-import com.sobok.deliveryservice.delivery.repository.RiderRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -39,72 +37,11 @@ import java.util.stream.Collectors;
 @Slf4j
 public class DeliveryService {
 
-    private final RiderRepository riderRepository;
     private final DeliveryRepository deliveryRepository;
-    private final AuthFeignClient authFeignClient;
     private final ShopFeignClient shopFeignClient;
     private final PaymentFeignClient paymentFeignClient;
     private final UserFeignClient userFeignClient;
-
-    public RiderResDto riderCreate(RiderReqDto dto) {
-
-        if (riderRepository.existsByPhone(dto.getPhone())) {
-            throw new CustomException("이미 사용 중인 전화번호입니다.", HttpStatus.BAD_REQUEST);
-        }
-        if (riderRepository.existsByPermissionNumber(dto.getPermissionNumber())) {
-            throw new CustomException("이미 사용 중인 면허번호입니다.", HttpStatus.BAD_REQUEST);
-        }
-
-        Rider rider = Rider.builder()
-                .authId(dto.getAuthId())
-                .name(dto.getName())
-                .phone(dto.getPhone())
-                .permissionNumber(dto.getPermissionNumber())
-                .build();
-
-        Rider saved = riderRepository.save(rider);
-
-        return RiderResDto.builder()
-                .id(saved.getId())
-                .authId(saved.getAuthId())
-                .name(saved.getName())
-                .phone(saved.getPhone())
-                .permissionNumber(saved.getPermissionNumber())
-                .build();
-    }
-
-    public ByPhoneResDto findByPhoneNumber(String phoneNumber) {
-        Optional<Rider> byPhone = riderRepository.findByPhone(phoneNumber);
-        if (byPhone.isPresent()) {
-            Rider rider = byPhone.get();
-            log.info("전화번호로 얻어온 rider 정보: {}", byPhone.toString());
-            return ByPhoneResDto.builder()
-                    .id(rider.getId())
-                    .authId(rider.getAuthId())
-                    .phone(rider.getPhone())
-                    .build();
-
-        } else {
-            log.info("해당 번호로 가입하신 정보가 없습니다.");
-            return null;
-        }
-    }
-
-    /**
-     * 라이더 정보 찾기
-     */
-    public AuthRiderInfoResDto getInfo(Long authId) {
-        Rider rider = riderRepository.getRiderByAuthId(authId).orElseThrow(
-                () -> new CustomException("해당하는 라이더가 존재하지 않습니다.", HttpStatus.BAD_REQUEST)
-        );
-
-        return AuthRiderInfoResDto.builder()
-                .phone(rider.getPhone())
-                .permissionNumber(rider.getPermissionNumber())
-                .name(rider.getName())
-                .loginId(null)
-                .build();
-    }
+    private final RiderService riderService;
 
     public void registerDelivery(DeliveryRegisterDto reqDto) {
         // 배달 객체 생성
@@ -116,36 +53,6 @@ public class DeliveryService {
         deliveryRepository.save(delivery);
     }
 
-    public Long getRiderId(Long id) {
-        Rider rider = riderRepository.getRiderByAuthId(id).orElseThrow(
-                () -> new CustomException("해당하는 라이더가 존재하지 않습니다.", HttpStatus.BAD_REQUEST)
-        );
-
-        return rider.getId();
-    }
-
-
-    /**
-     * 라이더 정보 조회
-     */
-    public List<RiderInfoResDto> getAllRiders() {
-        return riderRepository.findAll()
-                .stream()
-                .map(rider -> {
-                    RiderInfoResDto authInfo = authFeignClient.getRiderAuthInfo(rider.getAuthId());
-
-                    return RiderInfoResDto.builder()
-                            .id(rider.getId())
-                            .loginId(authInfo.getLoginId())
-                            .name(rider.getName())
-                            .phone(rider.getPhone())
-                            .permissionNumber(rider.getPermissionNumber())
-                            .active(authInfo.getActive())
-                            .build();
-                })
-                .toList();
-    }
-
     /**
      * 라이더 정보 조회 (전체 주문 조회용)
      */
@@ -153,8 +60,7 @@ public class DeliveryService {
         Delivery delivery = deliveryRepository.findByPaymentId(paymentId)
                 .orElseThrow(() -> new CustomException("배달 정보 없음", HttpStatus.NOT_FOUND));
 
-        Rider rider = riderRepository.findById(delivery.getRiderId())
-                .orElseThrow(() -> new CustomException("라이더 정보 없음", HttpStatus.NOT_FOUND));
+        Rider rider = riderService.findRiderByIdOrThrow(delivery.getRiderId());
 
         return RiderPaymentInfoResDto.builder()
                 .riderId(rider.getId())
@@ -163,22 +69,13 @@ public class DeliveryService {
                 .build();
     }
 
-    /**
-     * 라이더 면허번호 중복 체크
-     */
-    public void checkPermission(String permission) {
-        if (riderRepository.existsByPermissionNumber(permission)) {
-            throw new CustomException("사용할 수 없는 면허번호 입니다.", HttpStatus.BAD_REQUEST);
-        }
-    }
-
-
     public DeliveryResDto getDelivery(Long paymentId) {
-        Optional<Delivery> delivery = deliveryRepository.findByPaymentId(paymentId);
+        Delivery delivery = deliveryRepository.findByPaymentId(paymentId)
+                .orElseThrow(() -> new CustomException("해당 배달 목록이 존재하지 않습니다.", HttpStatus.NOT_FOUND));
         return DeliveryResDto.builder()
-                .shopId(delivery.get().getShopId())
-                .completeTime(delivery.get().getCompleteTime())
-                .riderId(delivery.get().getRiderId())
+                .shopId(delivery.getShopId())
+                .completeTime(delivery.getCompleteTime())
+                .riderId(delivery.getRiderId())
                 .build();
     }
 
@@ -199,9 +96,7 @@ public class DeliveryService {
             Long pageNo, Long numOfRows
     ) {
         //라이더 검증
-        if (!riderRepository.existsById(userInfo.getRiderId())) {
-            throw new CustomException("해당하는 라이더가 존재하지 않습니다.", HttpStatus.BAD_REQUEST);
-        }
+        riderService.existsByIdOrThrow(userInfo.getRiderId());
 
         //근처 가게 정보 목록 조회
         List<DeliveryAvailShopResDto> nearShop = shopFeignClient.getNearShop(latitude, longitude);
@@ -266,30 +161,6 @@ public class DeliveryService {
                 .toList();
     }
 
-    public ArrayList<RiderResDto> getPendingRiders() {
-        List<Long> inactiveRidersAuthIds;
-        try {
-            inactiveRidersAuthIds = authFeignClient.getInactiveRidersInfo();
-        } catch (Exception e) {
-            throw new CustomException("Feign 과정에서 오류가 발생했습니다.", HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-
-        if (inactiveRidersAuthIds.isEmpty()) {
-            return new ArrayList<>();
-        }
-
-        return riderRepository.findAll()
-                .stream()
-                .filter(rider -> inactiveRidersAuthIds.contains(rider.getAuthId()))
-                .map(rider -> RiderResDto.builder()
-                        .authId(rider.getAuthId())
-                        .name(rider.getName())
-                        .phone(rider.getPhone())
-                        .permissionNumber(rider.getPermissionNumber())
-                        .build())
-                .collect(Collectors.toCollection(ArrayList::new));
-    }
-
     /**
      * 배달 중인 목록 조회
      */
@@ -310,9 +181,7 @@ public class DeliveryService {
     // shop-service: deliveryList.getShopId로 가게 이름, 가게 주소 가져오기
     private List<DeliveryOrderResDto> getRiderDeliveries(TokenUserInfo userInfo, Long pageNo, Long numOfRows, boolean delivering) {
         // 1. 라이더 검증
-        if (!riderRepository.existsById(userInfo.getRiderId())) {
-            throw new CustomException("해당하는 라이더가 존재하지 않습니다.", HttpStatus.BAD_REQUEST);
-        }
+        riderService.existsByIdOrThrow(userInfo.getRiderId());
 
         // 2. 배달 목록 조회 (페이징 포함)
         Pageable pageable = PageRequest.of(pageNo.intValue() - 1, numOfRows.intValue(), Sort.by(Sort.Direction.DESC, "id"));
@@ -372,9 +241,7 @@ public class DeliveryService {
     @Transactional
     public void acceptDelivery(AcceptOrderReqDto acceptOrderReqDto) {
         //라이더 검증
-        if (!riderRepository.existsById(acceptOrderReqDto.getRiderId())) {
-            throw new CustomException("해당하는 라이더가 존재하지 않습니다.", HttpStatus.BAD_REQUEST);
-        }
+        riderService.existsByIdOrThrow(acceptOrderReqDto.getRiderId());
 
         //delivery 테이블에 riderId, completeTime 널인지 확인
         Delivery delivery = deliveryRepository.findByPaymentId(acceptOrderReqDto.getPaymentId())
@@ -395,9 +262,7 @@ public class DeliveryService {
     @Transactional
     public void deliveryComplete(AcceptOrderReqDto acceptOrderReqDto) {
         //라이더 검증
-        if (!riderRepository.existsById(acceptOrderReqDto.getRiderId())) {
-            throw new CustomException("해당하는 라이더가 존재하지 않습니다.", HttpStatus.BAD_REQUEST);
-        }
+        riderService.existsByIdOrThrow(acceptOrderReqDto.getRiderId());
 
         Delivery delivery = deliveryRepository.findByPaymentId(acceptOrderReqDto.getPaymentId())
                 .orElseThrow(() -> new CustomException("해당 배달 목록이 존재하지 않습니다.", HttpStatus.NOT_FOUND));
@@ -406,7 +271,7 @@ public class DeliveryService {
             throw new CustomException("이미 배달 완료된 주문입니다.", HttpStatus.BAD_REQUEST);
         }
 
-        if(!Objects.equals(acceptOrderReqDto.getRiderId(), delivery.getRiderId())) {
+        if (!Objects.equals(acceptOrderReqDto.getRiderId(), delivery.getRiderId())) {
             throw new CustomException("본인 담당 주문이 아닙니다.", HttpStatus.BAD_REQUEST);
         }
